pyproject.toml
[tool.poetry]
name = "wordle"
version = "0.1.0"
description = ""
authors = ["JulienGL <jgl94@free.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.2"
flask = "^3.1.0"
flask-cors = "^5.0.0"
python-dotenv = "^1.0.1"
types-flask-cors = "^5.0.0.20240902"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
